+++ tests/src/testmaterial.cpp
@@ -10,6 +10,9 @@
 #include <Eigen/Eigenvalues>
 
 #include <ikarus/finiteelements/mechanics/materials.hh>
+#if ENABLE_MUESLI
+  #include <ikarus/finiteelements/mechanics/materials/muesli/mueslimaterials.hh>
+#endif
 #include <ikarus/finiteelements/physicshelper.hh>
 #include <ikarus/utils/functionsanitychecks.hh>
 #include <ikarus/utils/init.hh>
@@ -135,10 +138,7 @@ auto testMaterialWithStrain(const MaterialImpl& mat, const double tol = 1e-13) {
 template <typename Material>
 auto testMaterial(Material mat) {
   TestSuite t("testMaterial");
-  if constexpr (std::is_same_v<Material, LinearElasticity> or
-                std::is_same_v<Material,
-                               Ikarus::VanishingStress<std::array<Ikarus::Impl::MatrixIndexPair, 1>({{{2, 2}}}),
-                                                       Ikarus::LinearElasticity>>) {
+  if constexpr (Material::isLinear) {
     t.subTest(testMaterialWithStrain<StrainTags::linear>(mat));
   } else if constexpr (std::is_same_v<Material,
                                       Ikarus::VanishingStress<std::array<Ikarus::Impl::MatrixIndexPair, 1>({{{2, 2}}}),
@@ -309,5 +309,37 @@ int main(int argc, char** argv) {
   t.subTest(testPlaneStrainAgainstPlaneStress<StrainTags::greenLagrangian, StVenantKirchhoff>());
   t.subTest(testPlaneStrainAgainstPlaneStress<StrainTags::rightCauchyGreenTensor, NeoHooke>());
 
+#if ENABLE_MUESLI
+  // Muesli
+  auto K  = convertLameConstants(matPar).toBulkModulus();
+  auto mu = convertLameConstants(matPar).toShearModulus();
+
+  auto muesliLin = Materials::makeMuesliLinearElasticity(matPar);
+  t.subTest(testMaterial(muesliLin));
+
+  auto muesliSVK = Materials::makeMuesliSVK(matPar);
+  t.subTest(testMaterial(muesliSVK));
+
+  auto muesliNeoHooke = Materials::makeMuesliNeoHooke(matPar, false);
+  t.subTest(testMaterial(muesliNeoHooke));
+
+  auto muesliNeoHookeReg = Materials::makeMuesliNeoHooke(matPar, true);
+  t.subTest(testMaterial(muesliNeoHookeReg));
+
+  auto muesliMR = Materials::makeMooneyRivlin({K, mu / 2, mu / 2});
+  t.subTest(testMaterial(muesliMR));
+
+  auto muesliYeoh = Materials::makeMuesliYeoh({mu / 2, mu / 6, mu / 3}, K);
+  t.subTest(testMaterial(muesliYeoh));
+
+  auto muesliAB = Materials::makeMuesliArrudaBoyce(mu, 0.85, K);
+  t.subTest(testMaterial(muesliAB));
+
+  auto muesliSVKPlaneStrain = planeStrain(muesliSVK);
+  t.subTest(testMaterial(muesliSVKPlaneStrain));
+
+  auto muesliNHPlaneStress = planeStress(muesliNeoHooke, 1e-12);
+  t.subTest(testMaterial(muesliNHPlaneStress));
+#endif
   return t.exit();
 }